backend/requirements.txt
Django==4.2.6
python-dotenv==1.0.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1

